+++ json-forms-compose/src/commonMain/kotlin/com/copperleaf/forms/compose/widgets/richtext/RichTextMarkdownSerializer.kt
@@ -1,0 +1,13 @@
+package com.copperleaf.forms.compose.widgets.richtext
+
+import com.darkrockstudios.richtexteditor.utils.RichTextValueSnapshot
+
+public class RichTextMarkdownSerializer : RichTextValueSnapshotSerializer {
+    override fun encodeToString(snapshot: RichTextValueSnapshot): String {
+        TODO("Not yet implemented")
+    }
+
+    override fun decodeFromString(content: String): RichTextValueSnapshot {
+        TODO("Not yet implemented")
+    }
+}