+++ json-forms-compose/src/commonMain/kotlin/com/copperleaf/forms/compose/widgets/richtext/RichTextValueSnapshotSerializer.kt
@@ -1,0 +1,9 @@
+package com.copperleaf.forms.compose.widgets.richtext
+
+import com.darkrockstudios.richtexteditor.utils.RichTextValueSnapshot
+
+public interface RichTextValueSnapshotSerializer {
+
+    public fun encodeToString(snapshot: RichTextValueSnapshot): String
+    public fun decodeFromString(content: String): RichTextValueSnapshot
+}