+++ json-forms-compose/src/commonTest/kotlin/com/copperleaf/forms/compose/widgets/richtext/TestRichTextHtmlSnapshot.kt
@@ -1,0 +1,4 @@
+package com.copperleaf.forms.compose.widgets.richtext
+
+class TestRichTextHtmlSnapshot {
+}