+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0 - 2022-07-05
+
+- Adds code editor control
+- Connects rich text widget to form state
+
 ## 0.2.0 - 2022-07-05
 
 - Adds a few more widget types